numcodecs_zfp_classic

Function compress

pub fn compress<T: ZfpCompressible, D: Dimension>(
    data: ArrayView<'_, T, D>,
    mode: &ZfpCompressionMode,
    non_finite: ZfpNonFiniteValuesMode,
) -> Result<Vec<u8>, ZfpClassicCodecError>

Compress the data array using ZFP with the provided mode.

Errors

Errors with

[ZfpCodecError::NonFiniteData] if any data element is non-finite (infinite or NaN) and a non-reversible lossy compression mode is used and the non_finite mode is not ZfpNonFiniteValuesMode::AllowUnsafe

ZfpClassicCodecError::ExcessiveDimensionality if data is more than 4-dimensional

ZfpClassicCodecError::InvalidExpertMode if the mode has invalid expert mode parameters

ZfpClassicCodecError::HeaderEncodeFailed if encoding the ZFP header failed

ZfpClassicCodecError::MetaHeaderEncodeFailed if encoding the array metadata header failed

ZfpClassicCodecError::ZfpEncodeFailed if an opaque encoding error occurred